+++ build.gradle
@@ -6,7 +6,7 @@ plugins {
 }
 
 group = 'com.sprint.mission'
-version = '2.0-M10'
+version = '2.2-M11'
 description = 'Demo project for Spring Boot'
 
 java {
