+++ src/main/java/com/sprint/mission/discodeit/dto/userDto/RoleUpdateRequest.java
@@ -1,6 +1,6 @@
 package com.sprint.mission.discodeit.dto.userDto;
 
-import com.sprint.mission.discodeit.entity.role.Role;
+import com.sprint.mission.discodeit.entity.enums.Role;
 
 import java.util.UUID;
