+++ src/main/java/com/sprint/mission/discodeit/entity/Channel.java
@@ -1,7 +1,7 @@
 package com.sprint.mission.discodeit.entity;
 
 import com.sprint.mission.discodeit.entity.base.BaseUpdatableEntity;
-import com.sprint.mission.discodeit.entity.role.ChannelType;
+import com.sprint.mission.discodeit.entity.enums.ChannelType;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
