+++ src/main/java/com/sprint/mission/discodeit/entity/enums/ChannelType.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity.role;
+package com.sprint.mission.discodeit.entity.enums;
 
 import lombok.Getter;
