+++ src/main/java/com/sprint/mission/discodeit/entity/enums/Role.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.entity.role;
+package com.sprint.mission.discodeit.entity.enums;
 
 public enum Role {
     ADMIN, CHANNEL_MANAGER, USER
