+++ src/main/java/com/sprint/mission/discodeit/security/jwt/JwtLoginSuccessHandler.java
@@ -1,7 +1,11 @@
-package com.sprint.mission.discodeit.security;
+package com.sprint.mission.discodeit.security.jwt;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.sprint.mission.discodeit.dto.jwt.JwtDto;
 import com.sprint.mission.discodeit.dto.userDto.UserDto;
+import com.sprint.mission.discodeit.security.CookieProvider;
+import com.sprint.mission.discodeit.security.DiscodeitUserDetails;
+import com.sprint.mission.discodeit.service.AuthService;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
@@ -16,9 +20,12 @@ import java.io.IOException;
 @Component
 @RequiredArgsConstructor
 @Slf4j
-public class LoginSuccessHandler implements AuthenticationSuccessHandler {
+public class JwtLoginSuccessHandler implements AuthenticationSuccessHandler {
 
     private final ObjectMapper objectMapper;
+    private final AuthService authService;
+//    private final JwtRegistry jwtRegistry;
+    private final CookieProvider cookieProvider;
 
     @Override
     public void onAuthenticationSuccess(HttpServletRequest request,
@@ -28,12 +35,15 @@ public class LoginSuccessHandler implements AuthenticationSuccessHandler {
         DiscodeitUserDetails userDetails = (DiscodeitUserDetails) authentication.getPrincipal();
         UserDto userDto = userDetails.getUserDto();
 
+        JwtDto jwtDto = authService.rotateRefreshToken(userDto);
+
+        cookieProvider.setRefreshTokenCookie(request, response, jwtDto.refreshToken());
+
         response.setStatus(HttpServletResponse.SC_OK);
         response.setContentType("application/json");
         response.setCharacterEncoding("UTF-8");
 
-        objectMapper.writeValue(response.getWriter(), userDto);
-
+        objectMapper.writeValue(response.getWriter(), jwtDto);
         log.info("로그인 성공: {}", userDto.username());
     }
 }
